added infer of model in ipynb

+++ vis_scene_colored2.py
@@ -12,7 +12,6 @@ from datasets import S3dDataset
 from pointnet import PointNetSeg
 
 
-import matplotlib.pyplot as plt
 from matplotlib import cm
 import open3d
 
@@ -20,12 +19,12 @@ from utils import get_s3d_cat2num
 
 
 def vis_numpy(points):
-    classes = np.unique(points[:, -1])
+    classes = np.unique(points[:, -1]).astype(np.int32)
     cmap = cm.get_cmap('jet')
 
     pcds = []
     for i, class_num in enumerate(classes):
-        classed_points = points[points[:, -1]==class_num]
+        classed_points = points[points[:, -1]==class_num].astype(np.float)
         colors = np.zeros((classed_points.shape[0], 3))
         colors[:] = cmap(i / len(classes))[:-1]
         pcd = open3d.geometry.PointCloud()
